+++ C64Studio/Debugger/VICERemoteDebuggerBinaryInterface.cs
@@ -818,7 +818,7 @@ namespace C64Studio
         if ( breakPoint.RemoteIndex == m_BrokenAtBreakPoint )
         {
           brokenBP = breakPoint;
-          breakAddress = breakPoint.Address;
+          breakAddress = (int)breakPoint.Address;
           if ( breakPoint.Temporary )
           {
             Log( "Remove auto startup breakpoint " + breakPoint.RemoteIndex );
@@ -1444,7 +1444,7 @@ namespace C64Studio
         if ( ( client != null )
         &&   ( client.Connected ) )
         {
-          RequestData requData = new RequestData( DebugRequestType.ADD_BREAKPOINT, BreakPoint.Address );
+          RequestData requData = new RequestData( DebugRequestType.ADD_BREAKPOINT, (int)BreakPoint.Address );
           requData.Breakpoint = BreakPoint;
           QueueRequest( requData );
         }
